server/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
python-multipart
pydantic
requests
